Fix up CS.

+++ src/Graph/Adapter/PhpDocumentorGraphAdapter.php
@@ -182,7 +182,7 @@ class PhpDocumentorGraphAdapter implements GraphAdapterInterface
      */
     protected function generateRandomString(int $length = 32): string
     {
-        $tokenLength = $length / 2;
+        $tokenLength = (int)($length / 2);
         $token = bin2hex(random_bytes($tokenLength));
 
         if (strlen($token) !== $length) {
